requirements.txt
streamlit>=1.20.0
sentence-transformers>=2.2.2
scikit-learn>=1.2.0
numpy>=1.23.0
pypdf>=3.10.0
python-dotenv>=1.0.0
tqdm>=4.65.0
pandas>=1.5.0
rapidfuzz>=3.0.0
